I fixed jumping with the scale controller :D
yipe
